Fix clippy issues in tests (#20)
Running `cargo clippy --tests` revealed several issues. This PR
addresses those issues.

Of note is that the `style` lint was changed to `deny`. This is because
clippy internally overrides a false positive of `needless_return` which
conflicts with `forbid`. The generated error message notes this will
soon become a hard error and links to [this
issue](https://github.com/rust-lang/rust/issues/81670) recommending
replacing `forbid` with `deny`.